Size

+++ 0x00-hello_world/6-size.c
@@ -1,5 +1,3 @@
-
-
 #include <stdio.h>
 /**
  * main -Entry point
@@ -13,10 +11,10 @@ long long int c;
 char d;
 float e;
 {
-printf("size of a char: %lu byte(s)\n", sizeof(d));
-printf("size of an int: %lu byte(s)\n", sizeof(a));
-printf("size of a long int: %lu byte(s)\n", sizeof(b));
-printf("size of a long long int: %lu byte(s)\n", sizeof(c));
-printf("size of a float: %lu byte(s)\n", sizeof(e));
+printf("size of a char: %lu byte(s)\n", (long unsigned int)sizeof(d));
+printf("size of an int: %lu byte(s)\n", (long unsigned int)sizeof(a));
+printf("size of a long int: %lu byte(s)\n", (long unsigned int)sizeof(b));
+printf("size of a long long int: %lu byte(s)\n", (long unsigned int)sizeof(c));
+printf("size of a float: %lu byte(s)\n", (long unsigned int)sizeof(e));
 return (0);
 }
